Add user login + invite functionality
Adds the ability for a user to log into the site and view pages
that require authentication. Adds the ability for an admin to send
invitation emails to potential users with a token-ized link. Adds the
ability for a potential user to visit the token-ized link and create
a username and password.

Modifies the layout template to allow for login/logout links to be
viewed at the appropriate time. Adds an admin dashboard to manage users
and add users.